+++ witdim_backend/main.go
@@ -2,6 +2,7 @@ package main
 
 import (
 	"github.com/ericarthurc/witdim/witdim_backend/database"
+	"github.com/ericarthurc/witdim/witdim_backend/models"
 	"github.com/ericarthurc/witdim/witdim_backend/routes"
 
 	"github.com/gofiber/fiber/v2"
@@ -12,6 +13,7 @@ func main() {
 	godotenv.Load("./.env")
 
 	database.ConnectDb()
+	models.CreateItemSchema()
 
 	app := fiber.New(fiber.Config{
 		// Prefork:      true,