+++ witdim_backend/routes/items.go
@@ -8,6 +8,7 @@ import (
 
 func ItemRoutes(app *fiber.App) {
 	items := app.Group("/api/v1/items")
+	items.Get("/search", handlers.GetItemsSearchHandler)
 	items.Get("/", handlers.GetItemsHandler)
 	items.Get("/:id", handlers.GetItemHandler)
 	items.Post("/", handlers.CreateItemHandler)