+++ witdim_frontend-web/src/App.tsx
@@ -1,7 +1,6 @@
 import { BrowserRouter as Router, Switch, Route, Link } from "react-router-dom";
 
 import Navbar from "./components/Navbar/Navbar";
-import Search from "./components/Search/Search";
 import Content from "./components/Content/Content";
 
 const App = () => {
@@ -10,7 +9,6 @@ const App = () => {
       <Navbar />
       {/* Placeholder Title LoL */}
       <h1 style={{ textAlign: "center" }}>WitDim</h1>
-      <Search />
       <Content />
     </Router>
   );