+++ witdim_backend/CHANGELOG.md
@@ -7,6 +7,13 @@ Addition of markdown posts will not be versioned or require a changelog update.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.2] - 2021-09-28
+
+## Added
+
+- `items/search` route and handler for searching the database
+- `text` index added to all fields for search
+
 ## [0.0.1] - 2021-09-28
 
 - Init commit